// theme/testing/fixtures.d.ts
export declare const utilityTestTheme: {
    elevate: {
        0: string;
        1: string;
        2: string;
        3: string;
        4: string;
        5: string;
    };
    radius: {
        circle: string;
        none: string;
        round: string;
        rounded: string;
    };
    unit: {
        0: string;
        1: string;
        2: string;
        3: string;
        4: string;
        5: string;
        6: string;
        7: string;
        8: string;
        9: string;
        10: string;
        default: string;
    };
};